Перенес combobox в редактор цепи из калькулятора
